refactor(FE): `useAuth` hook
