+++ frontend/src/components/Gym/Gym.js
@@ -1,11 +1,11 @@
 import React from "react";
 import { Link, useParams } from "react-router-dom";
-import { useAuth } from "../../context/auth-context";
 import { useFakemonsByGym, useGyms } from "../../hooks/queries";
 import ImportWalletAlert from "../ImportWalletAlert";
 import Button from "react-bootstrap/Button";
 import Card from "react-bootstrap/Card";
 import Alert from "../Alert";
+import useAuth from "../../hooks/useAuth";
 
 function Gym() {
   const { id: gymId } = useParams();
